+++ HISTORY.md
@@ -6,6 +6,7 @@
 * Fix `normalize_headers` not accepting header values in bytes
 * Fix inconsistency due to rounding in `CachedResponse.expires_unix` property
 * Fix form boundary used for cached multipart requests to _fully_ comply with RFC 2046
+* Fix loading cached JSON content with `decode_content=True` when the root element is an empty list or object
 * Fix usage example with `responses` library to be compatible with `requests` 2.32
 
 ## 1.2.0 (2024-02-17)